Refactor search bar and styles: Add SearchBar widget and stylesheet definitions
