+++ views/search_bar.py
@@ -1,0 +1,15 @@
+"""
+Module containing the SearchBar widget for the IPTV Player.
+"""
+
+from PyQt6.QtWidgets import QLineEdit
+from utils.styles import SearchBarStyle
+
+
+class SearchBar(QLineEdit):
+    """A custom search bar widget with styled appearance."""
+
+    def __init__(self):
+        super().__init__()
+        self.setPlaceholderText("🔍 Search channels...")
+        self.setStyleSheet(SearchBarStyle.SEARCH_BAR)
